+++ ibpy_native/interfaces/bridge.py
@@ -47,6 +47,27 @@ class IBridge(metaclass=abc.ABCMeta):
     ):
         pass
 
+    #region - Properties
+    @property
+    @abc.abstractmethod
+    def host(self) -> str:
+        """str: Hostname / IP address of the TWS / IB Gateway instance
+        connecting to.
+        """
+        return NotImplemented
+
+    @property
+    @abc.abstractmethod
+    def port(self) -> int:
+        """int: Socket port of this instance connecting to."""
+        return NotImplemented
+
+    @property
+    @abc.abstractmethod
+    def client_id(self) -> int:
+        """int: Client ID specified for this instance."""
+        return NotImplemented
+
     @property
     @abc.abstractmethod
     def is_connected(self) -> bool:
@@ -63,6 +84,7 @@ class IBridge(metaclass=abc.ABCMeta):
         events.
         """
         return NotImplemented
+    #endregion - Properties
 
     @abc.abstractmethod
     def set_timezone(self, tz: datetime.tzinfo):